requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
pillow>=9.5.0
numpy>=1.24.0
matplotlib>=3.7.0
opencv-python>=4.8.0
accelerate>=0.20.0
xformers>=0.0.20
datasets>=2.14.0
wandb>=0.15.0